requirements.txt
playwright
pandas
openpyxl
pywin32
rich
python-dotenv